Moved check_common_timeslot from Scheduler.py to Main.py
